Components and modules structure is changed.
