+++ src/app/app.module.ts
@@ -2,10 +2,10 @@ import {BrowserModule} from '@angular/platform-browser';
 import {NgModule} from '@angular/core';
 
 import {AppComponent} from './app.component';
-import {HeaderComponent} from './header/header.component';
-import {FooterComponent} from './footer/footer.component';
+import {HeaderComponent} from './core-module/components/header/header.component';
+import {FooterComponent} from './core-module/components/footer/footer.component';
 import {RouterModule} from '@angular/router';
-import {CoreModule} from './core-module/core.module';
+import {CoursesModule} from './courses-module/courses.module';
 import {AuthModule} from './auth-module/auth.module';
 import {SharedModule} from './shared-module/shared.module';
 
@@ -22,7 +22,7 @@ const routes = [
   imports: [
     BrowserModule,
     RouterModule.forRoot(routes),
-    CoreModule,
+    CoursesModule,
     AuthModule,
     SharedModule,
   ],
