+++ src/app/auth-module/components/login/login.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import {Router} from '@angular/router';
-import {Alert, AuthService} from '../services/auth.service';
+import {Alert, AuthService} from '../../services/auth.service';
 
 @Component({
   selector: 'app-login',
