+++ src/app/auth-module/components/registration/registration.component.spec.ts
