+++ src/app/core-module/components/footer/footer.component.css
@@ -1,7 +1,7 @@
 #footer {
   height: 64px;
   background-color: #414452;
-  position: absolute;
+  position: fixed;
   bottom: 0;
   width: 100%;
 }
