+++ src/app/core-module/components/footer/footer.component.ts
