+++ src/app/core-module/core.module.ts
@@ -1,39 +1,25 @@
-import {NgModule} from '@angular/core';
-import {CommonModule} from '@angular/common';
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import {HeaderComponent} from './components/header/header.component';
+import {FooterComponent} from './components/footer/footer.component';
+import {AuthModule} from '../auth-module/auth.module';
 import {RouterModule} from '@angular/router';
-import {CoursesComponent} from './courses/courses.component';
-import {CreateCourseComponent} from './create-course/create-course.component';
-import {SharedModule} from '../shared-module/shared.module';
-import {CourseComponent} from './course/course.component';
-import {ActionsComponent} from './actions/actions.component';
-import {FormsModule} from '@angular/forms';
-import {SearchComponent} from './search/search.component';
-import {LoadMoreCoursesComponent} from './load-more-courses/load-more-courses.component';
 
-const routes = [
-  {path: 'courses', component: CoursesComponent},
-  {path: 'create-course', component: CreateCourseComponent},
-];
+
 
 @NgModule({
   declarations: [
-    CourseComponent,
-    CoursesComponent,
-    CreateCourseComponent,
-    ActionsComponent,
-    SearchComponent,
-    LoadMoreCoursesComponent,
+    HeaderComponent,
+    FooterComponent,
   ],
   imports: [
     CommonModule,
-    RouterModule.forRoot(routes),
-    FormsModule,
-    SharedModule,
+    AuthModule,
+    RouterModule
   ],
   exports: [
-    CoursesComponent,
-    CreateCourseComponent,
+    HeaderComponent,
+    FooterComponent,
   ]
 })
-export class CoreModule {
-}
+export class CoreModule { }
