+++ src/app/courses-module/components/actions/actions.component.ts
