+++ src/app/courses-module/components/course/course.component.css
