+++ src/app/courses-module/components/load-more-courses/load-more-courses.component.css
@@ -11,3 +11,13 @@
 .disabled {
   display: none;
 }
+
+
+.test {
+  display: flex;
+}
+.test a {
+  vertical-align: center;
+  justify-content: center;
+/* left right*/
+}
