+++ src/app/courses-module/components/search/search.component.css
