+++ src/app/courses-module/models/course.ts
