+++ src/app/courses-module/services/course.service.ts
@@ -1,5 +1,5 @@
 import {Injectable} from '@angular/core';
-import {Course} from '../../entities/course';
+import {Course} from '../models/course';
 
 @Injectable({
   providedIn: 'root'
