+++ src/app/shared-module/shared.module.ts
@@ -1,6 +1,6 @@
 import {NgModule} from '@angular/core';
 import {CommonModule} from '@angular/common';
-import {BreadcrumbComponent} from './breadcrumb/breadcrumb.component';
+import {BreadcrumbComponent} from './components/breadcrumb/breadcrumb.component';
 import {RouterModule} from '@angular/router';
 
 @NgModule({
